Changed target_duration to float to support properly support sub-second durations. Added min_target_duration flag to hls, dash and segment muxers to support players which stop on a 0 target duration

+++ libavformat/hlsplaylist.c
@@ -72,7 +72,7 @@ void ff_hls_write_stream_info(AVStream *st, AVIOContext *out,
 }
 
 void ff_hls_write_playlist_header(AVIOContext *out, int version, int allowcache,
-                                  int target_duration, int64_t sequence,
+                                  float target_duration, int64_t sequence,
                                   uint32_t playlist_type) {
     if (!out)
         return;
